+++ iblsorter/__init__.py
@@ -2,8 +2,8 @@ import logging
 import os
 
 if os.getenv('MOCK_CUPY', False):
-    from pykilosort.testing.mock_cupy import cupy 
-    from pykilosort.testing.mock_cupyx import cupyx
+    from iblsorter.testing.mock_cupy import cupy
+    from iblsorter.testing.mock_cupyx import cupyx
 else:
     import cupy
     import cupyx
@@ -12,9 +12,7 @@ from .utils import Bunch, memmap_binary_file, read_data, load_probe, plot_dissim
 from .main import run, run_export, run_spikesort, run_preprocess
 from .io.probes import np1_probe, np2_probe, np2_4shank_probe
 
-
-__version__ = '1.5.0'
-
+__version__ = '1.7.3'
 
 # Set a null handler on the root logger
 logger = logging.getLogger(__name__)